Clean up ts setup
