+++ src/analyzer/MetadataDB.server.ts
@@ -1,7 +1,7 @@
 import { resolve } from "path"
 import os from "os"
 import { promises as fs } from "fs"
-import type { CompletedResult } from "./model"
+import type { CompletedResult } from "../shared/model"
 
 interface MetadataJson {
   completions: Record<string, { hash: string; time: number }>
