+++ src/analyzer/ServerInstance.server.ts
@@ -10,13 +10,13 @@ import type {
   FileModification,
   RenameInterval,
   FullCommitDTO
-} from "./model"
+} from "../shared/model"
 import { log } from "./log.server"
-import { analyzeRenamedFile } from "./util.server"
-import { contribRegex, gitLogRegex, gitLogRegexSimple, modeRegex, treeRegex } from "./constants"
+import { analyzeRenamedFile } from "~/shared/util"
+import { contribRegex, gitLogRegex, gitLogRegexSimple, modeRegex, treeRegex } from "../shared/constants"
 import { cpus, freemem, totalmem } from "node:os"
-import type { RepoData } from "~/routes/$repo.$"
-import type { InvocationReason } from "./RefreshPolicy"
+import type { RepoData } from "~/shared/model"
+import type { InvocationReason } from "../shared/RefreshPolicy"
 import InstanceManager from "./InstanceManager.server"
 
 export type AnalyzationStatus = "Starting" | "Hydrating" | "GeneratingChart"
