+++ src/analyzer/git-caller.server.ts
@@ -1,10 +1,13 @@
 import { log } from "./log.server"
-import { describeAsyncJob, getBaseDirFromPath, getDirName, promiseHelper, runProcess } from "./util.server"
+import { describeAsyncJob, runProcess } from "../shared/util.server"
+import { promiseHelper } from "~/shared/util"
+import { getBaseDirFromPath } from "~/shared/util.server"
+import { getDirName } from "~/shared/util.server"
 import { resolve, join } from "node:path"
 import { promises as fs, existsSync } from "node:fs"
-import type { AnalyzerData, GitRefs, Repository } from "./model"
-import { AnalyzerDataInterfaceVersion } from "./model"
-import { branchCompare, semverCompare } from "../util"
+import type { AnalyzerData, GitRefs, Repository } from "../shared/model"
+import { AnalyzerDataInterfaceVersion } from "../shared/model"
+import { branchCompare, semverCompare } from "../shared/util"
 import os from "node:os"
 import ServerInstance from "./ServerInstance.server"
