+++ src/cli.ts
@@ -5,9 +5,11 @@ import pkg from "../package.json"
 import getPort, { portNumbers } from "get-port"
 import open from "open"
 import { GitCaller } from "./analyzer/git-caller.server.ts"
-import { getArgsWithDefaults, parseArgs } from "./analyzer/args.server.ts"
-import { getPathFromRepoAndHead, generateVersionComparisonLink, semverCompare } from "./util.ts"
-import { describeAsyncJob, getDirName, getLatestVersion, isValidURI, promiseHelper } from "./analyzer/util.server.ts"
+import { getArgsWithDefaults, parseArgs } from "./shared/util.server.ts"
+import { getPathFromRepoAndHead, generateVersionComparisonLink, semverCompare } from "./shared/util.ts"
+import { describeAsyncJob, getLatestVersion } from "./shared/util.server.ts"
+import { isValidURI, promiseHelper } from "./shared/util.ts"
+import { getDirName } from "./shared/util.server.ts"
 import { log, setLogLevel } from "./analyzer/log.server.ts"
 import InstanceManager from "./analyzer/InstanceManager.server.ts"
