+++ src/components/Chart.tsx
@@ -3,7 +3,7 @@ import { hierarchy, pack, treemap, treemapResquarify } from "d3-hierarchy"
 import type { MouseEventHandler } from "react"
 import type { JSX } from "react"
 import { useDeferredValue, memo, useEffect, useMemo } from "react"
-import type { GitBlobObject, GitObject, GitTreeObject } from "~/analyzer/model"
+import type { GitBlobObject, GitObject, GitTreeObject } from "~/shared/model"
 import { useClickedObject } from "~/contexts/ClickedContext"
 import { useComponentSize } from "~/hooks"
 import {
@@ -25,11 +25,11 @@ import { useMetrics } from "../contexts/MetricContext"
 import type { ChartType } from "../contexts/OptionsContext"
 import { useOptions } from "../contexts/OptionsContext"
 import { usePath } from "../contexts/PathContext"
-import { getTextColorFromBackground, isBlob, isTree } from "~/util"
+import { getTextColorFromBackground, isBlob, isTree } from "~/shared/util"
 import clsx from "clsx"
 import type { SizeMetricType } from "~/metrics/sizeMetric"
 import { useSearch } from "~/contexts/SearchContext"
-import type { DatabaseInfo } from "~/routes/$repo.$"
+import type { DatabaseInfo } from "~/shared/model"
 import ignore, { type Ignore } from "ignore"
 import { cn, usePrefersLightMode } from "~/styling"
 import { isChrome, isChromium, isEdgeChromium } from "react-device-detect"
