+++ src/components/DetailsCard.tsx
@@ -5,7 +5,7 @@ import { type Fetcher, Form, useFetcher, useLocation, useNavigation } from "reac
 import byteSize from "byte-size"
 import clsx from "clsx"
 import { useEffect, useId, useMemo, useRef, useState } from "react"
-import type { GitObject, GitTreeObject } from "~/analyzer/model"
+import type { GitObject, GitTreeObject } from "~/shared/model"
 import { AuthorDistFragment } from "~/components/AuthorDistFragment"
 import { ChevronButton } from "~/components/ChevronButton"
 import { CloseButton } from "~/components/util"
@@ -15,7 +15,7 @@ import { useMetrics } from "~/contexts/MetricContext"
 import { useOptions } from "~/contexts/OptionsContext"
 import { usePath } from "~/contexts/PathContext"
 import { usePrefersLightMode } from "~/styling"
-import { dateFormatLong, getTextColorFromBackground, last } from "~/util"
+import { dateFormatLong, getTextColorFromBackground, last } from "~/shared/util"
 import { CommitsCard } from "./CommitsCard"
 import { MenuItem, MenuTab } from "./MenuTab"
