+++ src/components/Providers.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useMemo, useState } from "react"
-import type { GitBlobObject, GitObject } from "~/analyzer/model"
+import type { GitBlobObject, GitObject } from "~/shared/model"
 import { ClickedObjectContext } from "~/contexts/ClickedContext"
-import type { RepoData } from "~/routes/$repo.$"
+import type { RepoData } from "~/shared/model"
 import { DataContext } from "../contexts/DataContext"
 import { MetricsContext } from "../contexts/MetricContext"
 import type {
@@ -16,7 +16,7 @@ import { PathContext } from "../contexts/PathContext"
 import { SearchContext } from "../contexts/SearchContext"
 import type { MetricsData, MetricType } from "../metrics/metrics"
 import { createMetricData } from "../metrics/metrics"
-import { OPTIONS_LOCAL_STORAGE_KEY } from "~/analyzer/constants"
+import { OPTIONS_LOCAL_STORAGE_KEY } from "~/shared/constants"
 import type { SizeMetricType } from "~/metrics/sizeMetric"
 import type { DepthType } from "~/metrics/chartDepth"
 import type { CommitTab } from "~/contexts/CommitTabContext"
