+++ src/components/TimeSlider.tsx
@@ -2,7 +2,7 @@ import { useSubmit, useNavigation } from "react-router"
 import { useMemo, useState } from "react"
 import { Slider, Rail, Handles, Tracks } from "react-compound-slider"
 import { useData } from "~/contexts/DataContext"
-import { dateFormatCalendarHeader, dateFormatShort, getPathFromRepoAndHead } from "~/util"
+import { dateFormatCalendarHeader, dateFormatShort, getPathFromRepoAndHead } from "~/shared/util"
 import ClipLoader from "react-spinners/ClipLoader"
 import { Popover, ArrowContainer } from "react-tiny-popover"
 import DatePicker from "react-datepicker"
