+++ src/metrics/metrics.ts
@@ -1,11 +1,11 @@
 import sha1 from "sha1"
 import uniqolor from "uniqolor"
-import type { GitBlobObject, GitTreeObject } from "~/analyzer/model"
+import type { GitBlobObject, GitTreeObject } from "~/shared/model"
 import type { GradLegendData } from "~/components/legend/GradiantLegend"
 import type { PointInfo, PointLegendData } from "~/components/legend/PointLegend"
 import type { SegmentLegendData } from "~/components/legend/SegmentLegend"
 import { noEntryColor } from "~/const"
-import type { RepoData } from "~/routes/$repo.$"
+import type { RepoData } from "~/shared/model"
 import type { LegendType } from "../components/legend/Legend"
 import { setExtensionColor } from "./fileExtension"
 import { getLastChangedIndex, lastChangedGroupings } from "./lastChanged"
