+++ src/metrics/truckFactor.ts
@@ -1,6 +1,6 @@
-import type { GitBlobObject } from "~/analyzer/model"
+import type { GitBlobObject } from "~/shared/model"
 import type { MetricCache } from "./metrics"
-import { hslToHex } from "~/util"
+import { hslToHex } from "~/shared/util"
 import { noEntryColor } from "~/const"
 
 export class TruckFactorTranslater {
