+++ src/routes/$repo.$.tsx
@@ -1,6 +1,6 @@
 import { mdiChevronLeft, mdiChevronRight, mdiFullscreen, mdiFullscreenExit } from "@mdi/js"
 import Icon from "@mdi/react"
-import { redirect, Await, Link, isRouteErrorResponse, useLoaderData, useRouteError } from "react-router"
+import { Await, isRouteErrorResponse, useLoaderData, useRouteError, useLocation } from "react-router"
 import clsx from "clsx"
 import { resolve } from "path"
 import randomstring from "randomstring"
@@ -9,12 +9,11 @@ import { Suspense, memo, useEffect, useMemo, useRef, useState } from "react"
 import { Online } from "react-detect-offline"
 import { createPortal } from "react-dom"
 import { useMouse, useClient } from "~/hooks"
-import { getArgs } from "~/analyzer/args.server"
 import { GitCaller } from "~/analyzer/git-caller.server"
 import InstanceManager from "~/analyzer/InstanceManager.server"
-import type { CompletedResult, GitObject, GitTreeObject, Repository } from "~/analyzer/model"
-import { shouldUpdate } from "~/analyzer/RefreshPolicy"
-import { openFile } from "~/analyzer/util.server"
+import type { DatabaseInfo, GitObject, RepoData } from "~/shared/model"
+import { shouldUpdate } from "~/shared/RefreshPolicy"
+import { getArgs, openFile } from "~/shared/util.server"
 import BarChart from "~/components/BarChart"
 import { Breadcrumb } from "~/components/Breadcrumb"
 import { Chart } from "~/components/Chart"
@@ -30,65 +29,19 @@ import { SearchCard } from "~/components/SearchCard"
 import TimeSlider from "~/components/TimeSlider"
 import { Tooltip } from "~/components/Tooltip"
 import { UnionAuthorsModal } from "~/components/UnionAuthorsModal"
-import { ClearCacheForm, Code, ErrorPage } from "~/components/util"
+import { ErrorPage } from "~/components/util"
 
 import { cn } from "~/styling"
 import { log } from "~/analyzer/log.server"
 import type { Route } from "./+types/$repo.$"
+import { ClearCacheForm } from "./clearCache"
 
-export interface RepoData {
-  repo: Repository
-  gitTruckInfo: {
-    version: string
-    latestVersion: string | null
-  }
-  databaseInfo: DatabaseInfo
-}
+export const loader = async ({ params, context }: Route.LoaderArgs) => ({
+  dataPromise: analyze({ repo: params.repo, branch: params["*"] }),
+  versionInfo: context
+})
 
-export interface DatabaseInfo {
-  dominantAuthors: Record<string, { author: string; contribcount: number }>
-  commitCounts: Record<string, number>
-  lastChanged: Record<string, number>
-  authorCounts: Record<string, number>
-  maxCommitCount: number
-  minCommitCount: number
-  newestChangeDate: number
-  oldestChangeDate: number
-  authors: string[]
-  authorUnions: string[][]
-  fileTree: GitTreeObject
-  hiddenFiles: string[]
-  lastRunInfo: {
-    time: number
-    hash: string
-  }
-  fileCount: number
-  repo: string
-  branch: string
-  timerange: [number, number]
-  colorSeed: string | null
-  authorColors: Record<string, `#${string}`>
-  commitCountPerDay: { date: string; count: number }[]
-  selectedRange: [number, number]
-  analyzedRepos: CompletedResult[]
-  contribSumPerFile: Record<string, number>
-  maxMinContribCounts: { max: number; min: number }
-  commitCount: number
-}
-
-export const loader = async ({ params, context }: Route.LoaderArgs) => {
-  if (!params["repo"] || !params["*"]) {
-    return redirect("/")
-  }
-  const dataPromise = analyze(params, context)
-  return { dataPromise }
-}
-
-export const action = async ({ request, params }: Route.ActionArgs) => {
-  if (!params["repo"]) {
-    throw Error("This can never happen, since this route is only called if a repo exists in the URL")
-  }
-
+export const action = async ({ request, params: { repo, "*": branch } }: Route.ActionArgs) => {
   const formData = await request.formData()
   const refresh = formData.get("refresh")
   const unignore = formData.get("unignore")
@@ -101,8 +54,8 @@ export const action = async ({ request, params }: Route.ActionArgs) => {
   const authorcolor = formData.get("authorcolor")
 
   const args = await getArgs()
-  const path = resolve(args.path, params["repo"])
-  const instance = InstanceManager.getOrCreateInstance(params["repo"], params["*"] ?? "", path) // TODO fix the branch and check path works
+  const path = resolve(args.path, repo)
+  const instance = InstanceManager.getOrCreateInstance(repo, branch, path) // TODO fix the branch and check path works
   instance.prevInvokeReason = "unknown"
   if (refresh) {
     instance.prevInvokeReason = "refresh"
@@ -192,21 +145,19 @@ export const ErrorBoundary = () => {
   return <ErrorPage errorMessage={errorMessage} />
 }
 
-async function analyze(params: Route.LoaderArgs["params"], context: Route.LoaderArgs["context"]) {
+async function analyze({ repo, branch }: { repo: string; branch: string }) {
   const args = await getArgs()
-  const path = resolve(args.path, params["repo"] ?? "")
-  const branch = params["*"]
-  const repoName = params["repo"]
+  const path = resolve(args.path, repo)
   const isRepo = await GitCaller.isGitRepo(path)
   if (!isRepo) throw new Error(`No repo found at ${path}`)
-  if (!repoName || !branch) throw new Error(`Invalid repo and branch: ${repoName} ${branch}`)
   const isValidRevision = await GitCaller.isValidRevision(branch, path)
-  if (!isValidRevision)
+  if (!isValidRevision) {
     throw new Error(
-      `Invalid revision of repo ${params["repo"]}: ${branch}\nIf it is a remote branch, make sure it is pulled locally`
+      `Invalid revision of repo ${repo}: ${branch}\nIf ${branch} is a remote branch, make sure it is pulled locally`
     )
+  }
 
-  const instance = InstanceManager.getOrCreateInstance(repoName, branch, path)
+  const instance = InstanceManager.getOrCreateInstance(repo, branch, path)
   // to avoid double identical fetch at first load, which it does for some reason
   // TODO: Fix this. This is due to react strict mode
   if (instance.prevInvokeReason === "none" && instance.prevResult) {
@@ -217,9 +168,9 @@ async function analyze(params: Route.LoaderArgs["params"], context: Route.Loader
   const timerange = await instance.db.getOverallTimeRange()
   const selectedRange = instance.db.selectedRange
 
-  const repo = await GitCaller.getRepoMetadata(path)
+  const repositoryMetadata = await GitCaller.getRepoMetadata(path)
 
-  if (!repo) {
+  if (!repositoryMetadata) {
     throw Error("Error loading repo")
   }
 
@@ -326,18 +277,16 @@ async function analyze(params: Route.LoaderArgs["params"], context: Route.Loader
     commitCount
   }
 
-  const fullData = {
-    repo,
-    gitTruckInfo: context,
-    databaseInfo: databaseInfo
-  } as RepoData
+  const fullData: RepoData = { repo: repositoryMetadata, databaseInfo: databaseInfo }
 
   return fullData
 }
 
 export default function Repo() {
   const client = useClient()
-  const { dataPromise } = useLoaderData<typeof loader>()
+  const { dataPromise, versionInfo } = useLoaderData<typeof loader>()
+
+  const { pathname } = useLocation()
   const [isLeftPanelCollapse, setIsLeftPanelCollapse] = useState<boolean>(false)
   const [isRightPanelCollapse, setIsRightPanelCollapse] = useState<boolean>(false)
   const [isFullscreen, setIsFullscreen] = useState<boolean>(false)
@@ -375,7 +324,8 @@ export default function Repo() {
     <Suspense
       fallback={
         <div className="grid h-screen place-items-center">
-          <LoadingIndicator />
+          Stuck? Try clearing the cache:
+          <LoadingIndicator loadingText={<ClearCacheForm redirectPath={pathname} />} />
         </div>
       }
     >
@@ -384,31 +334,26 @@ export default function Repo() {
           <Providers data={data as RepoData}>
             <div className={cn("app-container", containerClass)}>
               <aside
-                className={clsx("grid auto-rows-min items-start gap-2 p-2 pr-0", {
-                  "overflow-y-auto": !isFullscreen
-                })}
+                className={clsx("grid auto-rows-min items-start gap-2 p-2 pr-0", { "overflow-y-auto": !isFullscreen })}
               >
                 {!isLeftPanelCollapse ? (
                   <>
-                    <GlobalInfo />
+                    <GlobalInfo
+                      installedVersion={versionInfo.installedVersion}
+                      latestVersion={versionInfo.latestVersion}
+                    />
                     <Options />
                     <Legend hoveredObject={hoveredObject} showUnionAuthorsModal={showUnionAuthorsModal} />
                   </>
                 ) : null}
                 {!isFullscreen ? (
-                  <div
-                    className={cn("absolute z-10 justify-self-end", {
-                      "left-0": isLeftPanelCollapse
-                    })}
-                  >
+                  <div className={cn("absolute z-10 justify-self-end", { "left-0": isLeftPanelCollapse })}>
                     <button
                       type="button"
                       onClick={() => setIsLeftPanelCollapse(!isLeftPanelCollapse)}
                       className={clsx(
                         "btn btn--primary absolute top-[50vh] left-0 flex h-6 w-6 cursor-pointer items-center justify-center rounded-full p-0",
-                        {
-                          "left-arrow-space": !isLeftPanelCollapse
-                        }
+                        { "left-arrow-space": !isLeftPanelCollapse }
                       )}
                     >
                       <Icon path={isLeftPanelCollapse ? mdiChevronRight : mdiChevronLeft} size={1} />
@@ -436,9 +381,7 @@ export default function Repo() {
               </main>
 
               <aside
-                className={clsx("grid auto-rows-min items-start gap-2 p-2 pl-0", {
-                  "overflow-y-auto": !isFullscreen
-                })}
+                className={clsx("grid auto-rows-min items-start gap-2 p-2 pl-0", { "overflow-y-auto": !isFullscreen })}
               >
                 {!isFullscreen ? (
                   <div className="absolute">
