+++ src/routes/$repository.tsx
@@ -1,10 +1,10 @@
 import { redirect } from "react-router"
 import { existsSync } from "node:fs"
 import { join } from "node:path"
-import { getArgsWithDefaults } from "~/analyzer/args.server"
+import { getArgsWithDefaults } from "~/shared/util.server"
 import { GitCaller } from "~/analyzer/git-caller.server"
 import type { Route } from "./+types/$repository"
-import { invariant } from "~/util"
+import { invariant } from "~/shared/util"
 
 /**
  * Redirect to the repository's default branch
