+++ src/shared/util.ts
@@ -1,7 +1,7 @@
 import type { HierarchyRectangularNode } from "d3-hierarchy"
 import { compare, valid, clean } from "semver"
 import colorConvert from "color-convert"
-import type { GitObject, GitBlobObject, GitTreeObject } from "./analyzer/model"
+import type { GitObject, GitBlobObject, GitTreeObject, RenameEntry } from "./model"
 
 export function diagonal(d: HierarchyRectangularNode<unknown>) {
   const dx = d.x1 - d.x0
@@ -61,7 +61,7 @@ export function dateFormatRelative(epochTime: number) {
   return "<1 hour ago"
 }
 
-export const last = <T>(arr: T[]) => arr[arr.length - 1]
+export const last = <T>(arr: T[]) => arr.at(-1)
 
 export const allExceptLast = <T>(arr: T[]) => {
   if (arr.length <= 1) return []
@@ -173,5 +173,125 @@ export function generateVersionComparisonLink({
 export function invariant<T>(condition: T, message: string): asserts condition is NonNullable<T> {
   if (!condition) {
     throw new Error(message)
+  }
+}
+
+export function sleep(ms: number) {
+  return new Promise((resolve) => {
+    setTimeout(resolve, ms)
+  })
+}
+export function getWeek(date: Date): number {
+  const tempDate = new Date(date)
+  tempDate.setHours(0, 0, 0, 0)
+  tempDate.setDate(tempDate.getDate() + 4 - (tempDate.getDay() || 7))
+  const yearStart = new Date(tempDate.getFullYear(), 0, 1)
+  const weekNo = Math.ceil(((tempDate.getTime() - yearStart.getTime()) / 86400000 + 1) / 7)
+  return weekNo
+}
+export function getTimeIntervals(timeUnit: string, minTime: number, maxTime: number): [string, number][] {
+  const intervals: [string, number][] = []
+
+  const startDate = new Date(minTime * 1000)
+  const endDate = new Date(maxTime * 1000)
+
+  const currentDate = new Date(startDate)
+
+  while (currentDate <= endDate) {
+    const currTime = currentDate.getTime() / 1000
+    if (timeUnit === "week") {
+      const weekNum = getWeek(currentDate)
+      intervals.push([`Week ${weekNum < 10 ? "0" : ""}${weekNum} ${currentDate.getFullYear()}`, currTime])
+      currentDate.setDate(currentDate.getDate() + 7)
+    } else if (timeUnit === "year") {
+      intervals.push([currentDate.getFullYear().toString(), currTime])
+      currentDate.setFullYear(currentDate.getFullYear() + 1)
+    } else if (timeUnit === "month") {
+      intervals.push([currentDate.toLocaleString("en-gb", { month: "long", year: "numeric" }), currTime])
+      currentDate.setMonth(currentDate.getMonth() + 1)
+    } else if (timeUnit === "day") {
+      intervals.push([
+        currentDate
+          .toLocaleDateString("en-gb", { day: "numeric", month: "long", year: "numeric", weekday: "short" })
+          .replace(",", ""),
+        currTime
+      ])
+      currentDate.setDate(currentDate.getDate() + 1)
+    }
+  }
+
+  return intervals
+}
+export function analyzeRenamedFile(
+  file: string,
+  timestamp: number,
+  authortime: number,
+  renamedFiles: RenameEntry[],
+  repo: string
+) {
+  const movedFileRegex = /(?:.*{(?<oldPath>.*)\s=>\s(?<newPath>.*)}.*)|(?:^(?<oldPath2>.*) => (?<newPath2>.*))$/gm
+  const replaceRegex = /{.*}/gm
+  const match = movedFileRegex.exec(file)
+  const groups = match?.groups ?? {}
+  let oldPath: string
+  let newPath: string
+
+  if (groups["oldPath"] || groups["newPath"]) {
+    const oldP = groups["oldPath"] ?? ""
+    const newP = groups["newPath"] ?? ""
+    oldPath = repo + "/" + file.replace(replaceRegex, oldP).replace("//", "/")
+    newPath = repo + "/" + file.replace(replaceRegex, newP).replace("//", "/")
+  } else {
+    oldPath = repo + "/" + (groups["oldPath2"] ?? "")
+    newPath = repo + "/" + (groups["newPath2"] ?? "")
+  }
+
+  renamedFiles.push({ fromname: oldPath, toname: newPath, timestamp: timestamp, timestampauthor: authortime })
+  return newPath
+}
+
+export function lookupFileInTree(tree: GitTreeObject, path: string): GitObject | undefined {
+  const dirs = path.split("/")
+
+  if (dirs.length < 2) {
+    // We have reached the end of the tree, look for the blob
+    const [file] = dirs
+    const result = tree.children.find((x) => x.name === file && x.type === "blob")
+    if (!result) return
+    return result
+  }
+  const subtree = tree.children.find((x) => x.name === dirs[0])
+  if (!subtree || subtree.type === "blob") return
+  return lookupFileInTree(subtree, dirs.slice(1).join("/"))
+}
+
+export const formatMs = (ms: number) => {
+  if (ms < 1000) {
+    return `${Math.round(ms)}ms`
+  } else {
+    return `${(ms / 1000).toFixed(2)}s`
+  }
+}
+
+/**
+ * This functions handles try / catch for you, so your code stays flat.
+ * @param promise An async function
+ * @returns A tuple of the result and an error. If there is no error, the error will be null.
+ */
+
+export async function promiseHelper<T>(promise: Promise<T>): Promise<[null, Error] | [T, null]> {
+  try {
+    return [await promise, null]
+  } catch (e) {
+    return [null, e as Error]
+  }
+}
+
+export function isValidURI(uri: string) {
+  try {
+    decodeURIComponent(uri)
+    return true
+  } catch (error) {
+    return false
   }
 }
